TimeParser interface, DataParser interface, удалил дублирование
