+++ src/main/java/ru/naumen/sd40/log/parser/ParsingMethod.java
@@ -1,0 +1,6 @@
+package ru.naumen.sd40.log.parser;
+
+public enum ParsingMethod {
+    SDNG,
+    GC
+}
