+++ src/main/java/ru/naumen/sd40/log/parser/TopParser.java
@@ -15,7 +15,7 @@ import java.util.regex.Pattern;
  * @author dkolmogortsev
  *
  */
-public class TopParser
+public class TopParser implements DataParser
 {
 
     private SimpleDateFormat sdf = new SimpleDateFormat("yyyyMMddHH:mm");
